website/requirements.txt
flask==2.3.0
requests==2.31.0
Werkzeug==2.3.0
gunicorn==21.2.0
flask-sqlalchemy==3.1.1
psycopg2-binary==2.9.9
psutil==5.9.6
stripe==7.4.0
twilio==8.10.1
oauthlib==3.2.2
flask-wtf==1.2.1
python-dotenv==1.0.0
email-validator==2.1.0.post1
Flask-Limiter==3.5.0
Flask-Caching==2.1.0
APScheduler==3.10.4 